Fix for issue #132 (Default font per OS)
Detect whether we're on XP, and if so, default to Verdana instead of
Arial.
From the bug, it looks like we only want to special case the default
font on XP. If we find later we want to do it per-OS, the code in
CheckOSForXP will need to be beefed up to return OS info instead of a
bool as it is currently.